update app to support newer versions of capybara
The method of invoking capybara servers changes with the 3.0 major
release, which is itself pretty old by now. Regardless, support that
change.

We run this gem off master, so I'm going to update the version, but I
can't just update the dev dependencies as we essentially aren't using
gem versioning at all.